creates packages for tictactoe
