+++ src/main/java/de/rabea/controller/tictactoe/TttHumanVsHumanController.java
@@ -1,19 +1,17 @@
-package de.rabea.controller;
+package de.rabea.controller.tictactoe;
 
 import de.rabea.Controller;
-import de.rabea.controller.html.BoardHtml;
-import de.rabea.controller.html.TicTacToeHtmlGenerator;
+import de.rabea.controller.tictactoe.html.BoardHtml;
+import de.rabea.controller.tictactoe.html.TicTacToeHtmlGenerator;
 import de.rabea.game.Board;
-import de.rabea.game.UnbeatableComputerPlayer;
 import de.rabea.request.HttpRequest;
 import de.rabea.request.MoveParser;
 import de.rabea.response.HttpResponse;
 
-import static de.rabea.game.GameMode.HumanVsComputer;
-import static de.rabea.game.Mark.O;
+import static de.rabea.game.GameMode.HumanVsHuman;
 import static de.rabea.response.head.StatusLine.OK;
 
-public class TttHumanVsComputerController extends Controller {
+public class TttHumanVsHumanController extends Controller {
 
     private Board board;
 
@@ -26,14 +24,11 @@ public class TttHumanVsComputerController extends Controller {
     @Override
     public HttpResponse doPost(HttpRequest request) {
         board = board.placeMark(new MoveParser(request.body()).move());
-        if (!board.gameOver()) {
-            board = board.placeMark(new UnbeatableComputerPlayer(O).getMove(board));
-        }
         return htmlBoard();
     }
 
     private HttpResponse htmlBoard() {
-        String html = new TicTacToeHtmlGenerator(new BoardHtml(board, HumanVsComputer)).generate();
+        String html = new TicTacToeHtmlGenerator(new BoardHtml(board, HumanVsHuman)).generate();
         return new HttpResponse(OK, html.getBytes());
     }
 }
