+++ src/test/java/de/rabea/controller/tictactoe/TicTacToeMenuControllerTest.java
@@ -1,5 +1,6 @@
-package de.rabea.controller;
+package de.rabea.controller.tictactoe;
 
+import de.rabea.controller.tictactoe.TicTacToeMenuController;
 import de.rabea.request.HttpRequest;
 import de.rabea.response.HttpResponse;
 import org.junit.Test;
@@ -38,7 +39,7 @@ public class TicTacToeMenuControllerTest {
     }
 
     @Test
-    public void redirectsToHvCGameUrlForPostRequestWithCvHChoice() {
+    public void redirectsToCvHGameUrlForPostRequestWithCvHChoice() {
         TicTacToeMenuController controller = new TicTacToeMenuController();
         HttpResponse response = controller.dispatch(new HttpRequest(POST, "/ttt-menu", ComputerVsHuman));
         String httpResponse = response.asString();
