+++ src/test/java/de/rabea/controller/tictactoe/TttHumanVsHumanControllerTest.java
@@ -1,5 +1,6 @@
-package de.rabea.controller;
+package de.rabea.controller.tictactoe;
 
+import de.rabea.controller.tictactoe.TttHumanVsHumanController;
 import de.rabea.request.HttpRequest;
 import de.rabea.response.HttpResponse;
 import org.junit.Test;
