+++ src/test/java/de/rabea/controller/tictactoe/html/TicTacToeHtmlGeneratorTest.java
@@ -1,5 +1,7 @@
-package de.rabea.controller.html;
+package de.rabea.controller.tictactoe.html;
 
+import de.rabea.controller.tictactoe.html.BoardHtml;
+import de.rabea.controller.tictactoe.html.TicTacToeHtmlGenerator;
 import de.rabea.game.Board;
 import org.junit.Test;
